remake dir structure proj
